feat: fakeRepo 분리, json 백업
