+++ app/models/requirements/fakeRepo.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+import { withSkillSlug } from '../skills/schema';
+import fakeRequirements from './fakeRequirements.json';
+import { requirementSchema } from './schema';
+
+export const fakeRequirementList = z
+  .array(withSkillSlug(requirementSchema))
+  .parse(fakeRequirements);
