+++ app/models/requirements/schema.ts
@@ -1,6 +1,4 @@
 import { z } from 'zod';
-import { withSkillSlug } from '../skills/schema';
-import fakeRequirements from './fakeRequirements.json';
 
 export const requirementSchema = z.object({
   rawText: z.string(),
@@ -9,6 +7,3 @@ export const requirementSchema = z.object({
 
 export type RequirementT = z.infer<typeof requirementSchema>;
 
-export const fakeRequirementList = z
-  .array(withSkillSlug(requirementSchema))
-  .parse(fakeRequirements);
