+++ app/models/skills/transformUtil.ts
@@ -5,5 +5,12 @@ export function flatMultiLink(transform: (item: any) => any){
   })
 }
 
+export function flatLink(transform: (item: any) => any){
+  return (name: string) => (item: any) => ({
+    ...item,
+    [name]: transform(item[name])
+  })
+}
+
 export const flatSlug = (e: any) => e.slug;
 export const flatSlugs = flatMultiLink(flatSlug)
